+++ sniffy-core/src/main/java/io/sniffy/SpyConfiguration.java
@@ -9,11 +9,17 @@ public class SpyConfiguration {
     private final boolean captureNetworkTraffic;
     private final boolean captureJdbc;
 
-    private SpyConfiguration(boolean captureStackTraces, boolean captureNetwork, boolean captureNetworkTraffic, boolean captureJdbc) {
+    /**
+     * @since 3.1.13
+     */
+    private final boolean bufferIncomingTraffic;
+
+    private SpyConfiguration(boolean captureStackTraces, boolean captureNetwork, boolean captureNetworkTraffic, boolean captureJdbc, boolean bufferIncomingTraffic) {
         this.captureStackTraces = captureStackTraces;
         this.captureNetwork = captureNetwork;
         this.captureNetworkTraffic = captureNetworkTraffic;
         this.captureJdbc = captureJdbc;
+        this.bufferIncomingTraffic = bufferIncomingTraffic;
     }
 
     public boolean isCaptureStackTraces() {
@@ -32,6 +38,13 @@ public class SpyConfiguration {
         return captureJdbc;
     }
 
+    /**
+     * @since 3.1.13
+     */
+    public boolean isBufferIncomingTraffic() {
+        return bufferIncomingTraffic;
+    }
+
     public static Builder builder() {
         return new Builder();
     }
@@ -42,10 +55,12 @@ public class SpyConfiguration {
         private boolean captureNetwork;
         private boolean captureNetworkTraffic;
         private boolean captureJdbc;
+        private boolean bufferIncomingTraffic;
 
         public Builder() {
             captureJdbc = SniffyConfiguration.INSTANCE.isMonitorJdbc();
             captureNetwork = SniffyConfiguration.INSTANCE.isMonitorSocket();
+            bufferIncomingTraffic = SniffyConfiguration.INSTANCE.isBufferIncomingTraffic();
             captureStackTraces = true;
         }
 
@@ -70,15 +85,24 @@ public class SpyConfiguration {
             return this;
         }
 
+        /**
+         * @since 3.1.13
+         */
+        public Builder bufferIncomingTraffic(boolean bufferIncomingTraffic) {
+            this.bufferIncomingTraffic = bufferIncomingTraffic;
+            return this;
+        }
+
         public Builder or(SpyConfiguration spyConfiguration) {
             return captureStackTraces(captureStackTraces || spyConfiguration.captureStackTraces).
                     captureNetwork(captureNetwork || spyConfiguration.captureNetwork).
                     captureNetworkTraffic(captureNetworkTraffic || spyConfiguration.captureNetworkTraffic).
-                    captureJdbc(captureJdbc || spyConfiguration.captureJdbc);
+                    captureJdbc(captureJdbc || spyConfiguration.captureJdbc).
+                    bufferIncomingTraffic(bufferIncomingTraffic || spyConfiguration.bufferIncomingTraffic);
         }
 
         public SpyConfiguration build() {
-            return new SpyConfiguration(captureStackTraces, captureNetwork, captureNetworkTraffic, captureJdbc);
+            return new SpyConfiguration(captureStackTraces, captureNetwork, captureNetworkTraffic, captureJdbc, bufferIncomingTraffic);
         }
 
     }