+++ sniffy-integration-tests/sniffy-integration-tests-decrypt-tls-bc/src/test/java/io/sniffy/tls/DecryptBouncyCastleGoogleTrafficTest.java
@@ -55,15 +55,24 @@ public class DecryptBouncyCastleGoogleTrafficTest {
                     URL url = new URL("https://www.google.com");
                     URLConnection urlConnection = url.openConnection();
 
+                    urlConnection.setRequestProperty("User-Agent", "Mozilla/5.0 (Macintosh; U; Intel Mac OS X 10.4; en-US; rv:1.9.2.2) Gecko/20100316 Firefox/3.6.2");
+
                     // On Java 14 with parallel builds sometimes throws SSLException: An established connection was aborted by the software in your host machine
                     //noinspection ResultOfMethodCallIgnored
                     urlConnection.getInputStream().read();
 
                     break;
                 } catch (IOException e) {
-                    if (e.getMessage().contains("An established connection was aborted by the software in your host machine") && OSUtil.isWindows() && (JVMUtil.getVersion() == 14 || JVMUtil.getVersion() == 13)) {
+                    if ((
+                            e.getMessage().contains("An established connection was aborted by the software in your host machine") ||
+                                    e.getMessage().contains("handshake_failure(40)")
+                    ) && OSUtil.isWindows() && (JVMUtil.getVersion() == 14 || JVMUtil.getVersion() == 13)) {
                         e.printStackTrace();
                         System.err.println("Caught " + e + " exception on Java " + JVMUtil.getVersion() + " running on Windows; retrying in 2 seconds");
+                        Thread.sleep(2000);
+                    } else if (e.getMessage().contains("Broken pipe") && OSUtil.isMac() && (JVMUtil.getVersion() >= 13)) {
+                        e.printStackTrace();
+                        System.err.println("Caught " + e + " exception on Java " + JVMUtil.getVersion() + " running on Mac OS; retrying in 2 seconds");
                         Thread.sleep(2000);
                     } else {
                         throw e;