+++ sniffy-integration-tests/sniffy-integration-tests-decrypt-tls/src/test/java/io/sniffy/tls/DecryptGoogleTrafficTest.java
@@ -48,7 +48,8 @@ public class DecryptGoogleTrafficTest {
 
                     break;
                 } catch (IOException e) {
-                    if (e.getMessage().contains("An established connection was aborted by the software in your host machine") && OSUtil.isWindows() && (JVMUtil.getVersion() == 14 || JVMUtil.getVersion() == 13)) {
+                    if (e.getMessage().contains("An established connection was aborted by the software in your host machine") || e.getMessage().contains("Remote host terminated the handshake")
+                            && OSUtil.isWindows() && (JVMUtil.getVersion() == 14 || JVMUtil.getVersion() == 13)) {
                         e.printStackTrace();
                         System.err.println("Caught " + e + " exception on Java " + JVMUtil.getVersion() + " running on Windows; retrying in 2 seconds");
                         Thread.sleep(2000);