+++ sniffy-module-nio-compat-deps/src/main/java-compat/java/net/ProtocolFamily.java
@@ -1,7 +1,0 @@
-package java.net;
-
-public interface ProtocolFamily {
-
-    String name();
-
-}