+++ sniffy-module-nio/src/main/java/io/sniffy/nio/SniffySelector.java
@@ -26,8 +26,11 @@ public class SniffySelector extends AbstractSelector {
 
     private final AbstractSelector delegate;
 
+    private volatile Set<SelectionKey> keysWrapper = null;
+    private volatile Set<SelectionKey> selectedKeysWrapper = null;
+
     private final Map<AbstractSelectableChannel, AbstractSelectableChannel> channelToSniffyChannelMap =
-            Collections.synchronizedMap(new WeakHashMap<AbstractSelectableChannel, AbstractSelectableChannel>());
+            new WeakHashMap<AbstractSelectableChannel, AbstractSelectableChannel>();
 
     private final Map<SelectionKey, SniffySelectionKey> sniffySelectionKeyCache =
             new WeakHashMap<SelectionKey, SniffySelectionKey>();
@@ -38,13 +41,13 @@ public class SniffySelector extends AbstractSelector {
         LOG.trace("Created new SniffySelector(" + provider + ", " + delegate + ") = " + this);
     }
 
-    public SniffySelectionKey wrap(SelectionKey delegate, SniffySelector sniffySelector, SelectableChannel sniffyChannel) {
+    public SniffySelectionKey wrap(SelectionKey delegate, SniffySelector sniffySelector, SelectableChannel sniffySocketChannel) {
         SniffySelectionKey sniffySelectionKey = sniffySelectionKeyCache.get(delegate);
         if (null == sniffySelectionKey) {
             synchronized (sniffySelectionKeyCache) {
                 sniffySelectionKey = sniffySelectionKeyCache.get(delegate);
                 if (null == sniffySelectionKey) {
-                    sniffySelectionKey = new SniffySelectionKey(delegate, sniffySelector, sniffyChannel);
+                    sniffySelectionKey = new SniffySelectionKey(delegate, sniffySelector, sniffySocketChannel);
                     sniffySelectionKeyCache.put(delegate, sniffySelectionKey);
                 }
             }
@@ -56,6 +59,7 @@ public class SniffySelector extends AbstractSelector {
     @Override
     protected void implCloseSelector() throws IOException {
         try {
+            // TODO: document
             setField(AbstractSelector.class, delegate, "closed", true);
             invokeMethod(AbstractSelector.class, delegate, "implCloseSelector", Void.class);
         } catch (Exception e) {
@@ -63,34 +67,48 @@ public class SniffySelector extends AbstractSelector {
         }
     }
 
-    private Set<SelectionKey> wrapSelectionKeys(final Set<SelectionKey> delegates) {
-        if (null == delegates) {
-            return null;
-        } else if (delegates.isEmpty()) {
-            return Collections.<SelectionKey>emptySet();
-        } else {
-            return new SetWrapper<SniffySelectionKey, SelectionKey>(delegates, new WrapperFactory<SelectionKey, SniffySelectionKey>() {
-
-                @Override
-                public SniffySelectionKey wrap(SelectionKey delegate) {
-                    //noinspection SuspiciousMethodCalls
-                    return SniffySelector.this.wrap(delegate, SniffySelector.this, channelToSniffyChannelMap.get(delegate.channel()));
+    private Set<SelectionKey> wrapKeys(final Set<SelectionKey> delegates) {
+        if (null == keysWrapper) {
+            synchronized (this) {
+                if (null == keysWrapper && null != delegates) {
+                    keysWrapper = createSelectionKeysWrapper(delegates);
                 }
-
-            });
+            }
         }
+        return keysWrapper;
+    }
+
+    private Set<SelectionKey> wrapSelectedKeys(final Set<SelectionKey> delegates) {
+        if (null == selectedKeysWrapper) {
+            synchronized (this) {
+                if (null == selectedKeysWrapper && null != delegates) {
+                    selectedKeysWrapper = createSelectionKeysWrapper(delegates);
+                }
+            }
+        }
+        return selectedKeysWrapper;
+    }
+
+    private SetWrapper<SniffySelectionKey, SelectionKey> createSelectionKeysWrapper(Set<SelectionKey> delegates) {
+        return new SetWrapper<SniffySelectionKey, SelectionKey>(delegates, new WrapperFactory<SelectionKey, SniffySelectionKey>() {
+            @Override
+            public SniffySelectionKey wrap(SelectionKey delegate) {
+                //noinspection SuspiciousMethodCalls
+                return SniffySelector.this.wrap(delegate, SniffySelector.this, channelToSniffyChannelMap.get(delegate.channel()));
+            }
+        });
     }
 
     private class SelectionKeyConsumerWrapper implements Consumer<SelectionKey> {
 
         private final Consumer<SelectionKey> delegate;
-        private final AbstractSelectableChannel sniffyChannel;
+        private final SniffySocketChannel sniffyChannel;
 
         public SelectionKeyConsumerWrapper(Consumer<SelectionKey> delegate) {
             this(delegate, null);
         }
 
-        public SelectionKeyConsumerWrapper(Consumer<SelectionKey> delegate, AbstractSelectableChannel sniffyChannel) {
+        public SelectionKeyConsumerWrapper(Consumer<SelectionKey> delegate, SniffySocketChannel sniffyChannel) {
             this.delegate = delegate;
             this.sniffyChannel = sniffyChannel;
         }
@@ -106,6 +124,7 @@ public class SniffySelector extends AbstractSelector {
      * This method adds a selection key to provided AbstractSelectableChannel, hence we're doing the same here manually
      */
     @Override
+    // TODO: document
     protected SelectionKey register(AbstractSelectableChannel ch, int ops, Object att) {
         try {
 
@@ -123,14 +142,17 @@ public class SniffySelector extends AbstractSelector {
                     SelectionKey.class
             );
 
+            Object regLock = ReflectionUtil.getField(AbstractSelectableChannel.class, chDelegate, "regLock");
             Object keyLock = ReflectionUtil.getField(AbstractSelectableChannel.class, chDelegate, "keyLock");
             //noinspection SynchronizationOnLocalVariableOrMethodParameter
-            synchronized (keyLock) {
-                invokeMethod(AbstractSelectableChannel.class, chDelegate, "addKey", SelectionKey.class, selectionKeyDelegate, Void.class);
+            synchronized (regLock) {
+                //noinspection SynchronizationOnLocalVariableOrMethodParameter
+                synchronized (keyLock) {
+                    invokeMethod(AbstractSelectableChannel.class, chDelegate, "addKey", SelectionKey.class, selectionKeyDelegate, Void.class);
+                }
             }
 
             return wrap(selectionKeyDelegate, this, ch);
-
 
         } catch (Exception e) {
             throw ExceptionUtil.processException(e);
@@ -139,12 +161,19 @@ public class SniffySelector extends AbstractSelector {
 
     @Override
     public Set<SelectionKey> keys() {
-        return wrapSelectionKeys(delegate.keys());
+        return wrapKeys(delegate.keys());
     }
 
     @Override
     public Set<SelectionKey> selectedKeys() {
-        return wrapSelectionKeys(delegate.selectedKeys());
+        return wrapSelectedKeys(delegate.selectedKeys());
+    }
+
+    private void copyKeysFromDelegate() {
+
+        // TODO: copy keys from delegate channels to sniffy channels
+        // TODO: consider doing in background?
+
     }
 
     /**
@@ -153,63 +182,7 @@ public class SniffySelector extends AbstractSelector {
      */
     @Override
     public int selectNow() throws IOException {
-        try {
-            return delegate.selectNow();
-        } finally {
-            updateSelectionKeysFromDelegate();
-        }
-    }
-
-    /**
-     * select method can remove cancelled selection keys from delegate so we need to update them in sniffy channels as well
-     */
-    private void updateSelectionKeysFromDelegate() {
-
-        if (JVMUtil.getVersion() < 14 && !Boolean.getBoolean("io.sniffy.forceJava14Compatibility")) {
-            return; // Before Java 14 is updating attachment in delegate from SniffySelectionKey
-        }
-
-        Map<AbstractSelectableChannel, AbstractSelectableChannel> channelToSniffyChannelMap;
-        synchronized (this.channelToSniffyChannelMap) {
-            channelToSniffyChannelMap = new HashMap<AbstractSelectableChannel, AbstractSelectableChannel>(this.channelToSniffyChannelMap);
-        }
-
-        for (Map.Entry<AbstractSelectableChannel, AbstractSelectableChannel> entry : channelToSniffyChannelMap.entrySet()) {
-            AbstractSelectableChannel sniffyChannel = entry.getValue();
-
-            if (sniffyChannel instanceof SelectableChannelWrapper) {
-                AbstractSelectableChannel delegate = ((SelectableChannelWrapper<?>) sniffyChannel).getDelegate();
-                try {
-
-                    Object keyLock = ReflectionUtil.getField(AbstractSelectableChannel.class, sniffyChannel, "keyLock");
-                    Object delegateKeyLock = ReflectionUtil.getField(AbstractSelectableChannel.class, delegate, "keyLock");
-
-                    // We're always obtaining monitor of wrapper (SniffySelector) first in order to avoid dead locks
-
-                    //noinspection SynchronizationOnLocalVariableOrMethodParameter
-                    synchronized (keyLock) {
-                        //noinspection SynchronizationOnLocalVariableOrMethodParameter
-                        synchronized (delegateKeyLock) {
-
-                            SelectionKey[] delegateKeys = ReflectionUtil.getField(AbstractSelectableChannel.class, delegate, "keys");
-                            List<SelectionKey> sniffyKeys = new ArrayList<SelectionKey>(delegateKeys.length);
-                            for (SelectionKey delegateKey : delegateKeys) {
-                                sniffyKeys.add(null == delegateKey ? null : wrap(delegateKey, this, sniffyChannel));
-                            }
-                            ReflectionUtil.setField(AbstractSelectableChannel.class, sniffyChannel, "keys", sniffyKeys.toArray(new SelectionKey[0]));
-
-                            ReflectionUtil.setField(AbstractSelectableChannel.class, delegate, "keyCount"
-                                    ,ReflectionUtil.getField(AbstractSelectableChannel.class, sniffyChannel, "keyCount"));
-
-                        }
-                    }
-
-                } catch (Exception e) {
-                    e.printStackTrace();
-                }
-            }
-        }
-
+        return delegate.selectNow();
     }
 
     /**
@@ -218,11 +191,7 @@ public class SniffySelector extends AbstractSelector {
      */
     @Override
     public int select(long timeout) throws IOException {
-        try {
-            return delegate.select(timeout);
-        } finally {
-            updateSelectionKeysFromDelegate();
-        }
+        return delegate.select(timeout);
     }
 
     /**
@@ -231,11 +200,7 @@ public class SniffySelector extends AbstractSelector {
      */
     @Override
     public int select() throws IOException {
-        try {
-            return delegate.select();
-        } finally {
-            updateSelectionKeysFromDelegate();
-        }
+        return delegate.select();
     }
 
     @Override
@@ -246,9 +211,10 @@ public class SniffySelector extends AbstractSelector {
 
     // Note: this method was absent in earlier JDKs so we cannot use @Override annotation
     //@Override
-    @SuppressWarnings("RedundantThrows")
+    @SuppressWarnings({"RedundantThrows", "Since15"})
     public int select(Consumer<SelectionKey> action, long timeout) throws IOException {
         try {
+            // TODO: call delegate.processDeregisterQueue and update selection keys from delegate
             return invokeMethod(Selector.class, delegate, "select",
                     Consumer.class, new SelectionKeyConsumerWrapper(action),
                     Long.TYPE, timeout,
@@ -256,14 +222,12 @@ public class SniffySelector extends AbstractSelector {
             );
         } catch (Exception e) {
             throw ExceptionUtil.processException(e);
-        } finally {
-            updateSelectionKeysFromDelegate();
         }
     }
 
     // Note: this method was absent in earlier JDKs so we cannot use @Override annotation
     //@Override
-    @SuppressWarnings("RedundantThrows")
+    @SuppressWarnings({"RedundantThrows", "Since15"})
     public int select(Consumer<SelectionKey> action) throws IOException {
         try {
             return invokeMethod(Selector.class, delegate, "select",
@@ -272,14 +236,12 @@ public class SniffySelector extends AbstractSelector {
             );
         } catch (Exception e) {
             throw ExceptionUtil.processException(e);
-        } finally {
-            updateSelectionKeysFromDelegate();
         }
     }
 
     // Note: this method was absent in earlier JDKs so we cannot use @Override annotation
     //@Override
-    @SuppressWarnings("RedundantThrows")
+    @SuppressWarnings({"RedundantThrows", "Since15"})
     public int selectNow(Consumer<SelectionKey> action) throws IOException {
         try {
             return invokeMethod(Selector.class, delegate, "selectNow",
@@ -288,8 +250,6 @@ public class SniffySelector extends AbstractSelector {
             );
         } catch (Exception e) {
             throw ExceptionUtil.processException(e);
-        } finally {
-            updateSelectionKeysFromDelegate();
         }
     }